Make WorkLookup and ItemLookup separate classes
